New folder (2)/search_project/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
transformers
sentence-transformers
faiss-cpu
jinja2
